+++ conans/model/__init__.py
@@ -1,0 +1,45 @@
+from abc import ABCMeta, abstractproperty
+
+
+class Generator(object):
+    __metaclass__ = ABCMeta
+
+    def __init__(self, conanfile):
+        self.conanfile = conanfile
+        self._deps_build_info = conanfile.deps_cpp_info
+        self._build_info = conanfile.cpp_info
+        self._deps_env_info = conanfile.deps_env_info
+        self._env_info = conanfile.env_info
+        self._deps_user_info = conanfile.deps_user_info
+
+    @property
+    def deps_build_info(self):
+        return self._deps_build_info
+
+    @property
+    def build_info(self):
+        return self._build_info
+
+    @property
+    def deps_env_info(self):
+        return self._deps_env_info
+
+    @property
+    def deps_user_info(self):
+        return self._deps_user_info
+
+    @property
+    def env_info(self):
+        return self._env_info
+
+    @property
+    def settings(self):
+        return self.conanfile.settings
+
+    @abstractproperty
+    def content(self):
+        raise NotImplementedError()
+
+    @abstractproperty
+    def filename(self):
+        raise NotImplementedError()
